# backend/src/docs/swagger.yaml
definitions:
...
  dto.ErrorCode:
    properties:
      errorCode:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        type: integer
        x-enum-comments:
          device_not_found: '"Device not found {{.DeviceID}}",404'
          generic_error: '"An unexpected error occurred",500'
          invalid_parameter: '"Invalid parameter: {{.Key}}. {{.Message}}",405'
          json_marshal_error: '"Unable to marshal JSON: {{.Error}}",500'
          json_unmarshal_error: '"Unable to unmarshal JSON: {{.Error}}",500'
          mount_fail: '"Unable to mount {{.Device}} on {{.Path}}. {{.Message}}",406'
          unknown: invalid
          unmount_fail: '"Unable to unmount {{.ID}}. {{.Message}}",406'
        x-enum-varnames:
        - unknown
        - generic_error
        - json_marshal_error
        - json_unmarshal_error
        - invalid_parameter
        - mount_fail
        - unmount_fail
        - device_not_found
      errorMessage:
        type: string
      httpCode:
        type: integer
    type: object
  dto.ErrorInfo:
    properties:
      code:
        $ref: '#/definitions/dto.ErrorCode'
      data:
        additionalProperties: {}
        type: object
      deep_message:
        type: string
      error: {}
      message:
        type: string
      trace:
        items:
          $ref: '#/definitions/tracerr.Frame'
        type: array
    type: object
  dto.EventMessageEnvelope:
    properties:
      data: {}
      event:
        $ref: '#/definitions/dto.EventType'
      id:
        type: string
    type: object
  dto.EventType:
    enum:
    - hello
    - update
    - heartbeat
    - share
    - volumes
    - dirty
    type: string
    x-enum-varnames:
    - EventHello
    - EventUpdate
    - EventHeartbeat
    - EventShare
    - EventVolumes
    - EventDirty
...
  dto.Settings:
    properties:
      allow_hosts:
        items:
          type: string
        type: array
      bind_all_interfaces:
        type: boolean
      compatibility_mode:
        type: boolean
      interfaces:
        items:
          type: string
        type: array
      log_level:
        type: string
      mountoptions:
        items:
          type: string
        type: array
      multi_channel:
        type: boolean
      recyle_bin_enabled:
        type: boolean
      update_channel:
        $ref: '#/definitions/dto.UpdateChannel'
      veto_files:
        items:
          type: string
        type: array
      workgroup:
        type: string
    type: object
...
  dto.UpdateChannel:
    enum:
    - stable
    - prerelease
    - none
    type: string
    x-enum-varnames:
    - Stable
    - Prerelease
    - None
...
  tracerr.Frame:
    properties:
      func:
        description: Func contains a function name.
        type: string
      line:
        description: Line contains a line number.
        type: integer
      path:
        description: Path contains a file path.
        type: string
    type: object
info:
  contact:
    email: lucio.tarantino@gmail.com
    name: Lucio Tarantino
    url: https://github.com/dianlight
  description: This are samba rest admin API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SRAT API
  version: "1.0"
paths:
...
  /settings:
    get:
      consumes:
      - application/json
      description: Get the configuration for the global samba settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
      summary: Get the configuration for the global samba settings
      tags:
      - samba
    patch:
      consumes:
      - application/json
      description: Update the configuration for the global samba settings
      parameters:
      - description: Update model
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
      summary: Update the configuration for the global samba settings
      tags:
      - samba
    put:
      consumes:
      - application/json
      description: Update the configuration for the global samba settings
      parameters:
      - description: Update model
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
      summary: Update the configuration for the global samba settings
      tags:
      - samba
...
  /sse:
    get:
      consumes:
      - application/json
      description: Open a SSE stream
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventMessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorInfo'
      summary: Open a SSE stream
      tags:
      - system
  /sse/events:
    get:
      description: Return a list of available WSChannel events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EventType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: EventTypeList
      tags:
      - system
...
securityDefinitions:
  ApiKeyAuth:
    description: HomeAssistant Supervisor Token
    in: header
    name: X-Supervisor-Token
    type: apiKey
swagger: "2.0"